Updating: (improving search_func in CommandApp class to search strict name), and adding (bpm as parameter in print_info function in InterfaceApp)

+++ src/midiplayer.py
@@ -1057,7 +1057,7 @@ class MidiBase(object):
         """
 
         # debug("sec_per_tick: {}".format(self.sec_per_tick))
-        return nb_tick * self.sec_per_tick
+        return float(nb_tick * self.sec_per_tick)
 
     #-----------------------------------------
